requirements.txt
# Core TTS
chatterbox-tts>=0.1.6

# Web server
flask>=3.0.0

# Audio processing (installed with chatterbox, listed for clarity)
torchaudio>=2.6.0
soundfile>=0.12.0

# Note: PyTorch with CUDA must be installed separately:
# pip install torch==2.6.0 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124
